Update Future Plan Page
Finished creating UI changes in the future plan page. All of the functionality of the page from before the changes has been maintained. The overall appearance of the page should be about complete. We may need to replace the red x's to delete classes eventually, since they look a tiny bit strange.